refactor: Application 클래스에서 불필요한 상태와 로직 제거
