+++ app/src/main/java/woowacourse/omok/StartActivity.kt
@@ -19,7 +19,7 @@ class StartActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_start)
 
-        nicknameView.setText(Omok.getNickname())
+        nicknameView.setText(Storage.getNickname())
 
         val startButton = findViewById<Button>(R.id.start_game)
         startButton.setOnClickListener {
@@ -80,7 +80,7 @@ class StartActivity : AppCompatActivity() {
 
     private fun goToGame(nickname: String) {
         val intent = Intent(this, MainActivity::class.java)
-        Omok.editNickname(nickname)
+        Storage.editNickname(nickname)
         startActivity(intent)
     }
